+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+* Add telescope.nvim with its configuration files.
+* Add nvim-treesitter.nvim with its configuration files.
 * Add the kanagawa.nvim color scheme.
 * Create CHANGELOG.md, add Lazy plugin manager, and update the README.md,